requirements.txt
Flask==2.2.2
Flask_WTF==1.1.1
Requests==2.31.0
Werkzeug==2.2.2
matplotlib==3.7.1
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==10.1.0
scipy==1.11.4